requirements.txt
attrs
certifi
chardet
Deprecated
idna
importlib-metadata
jsonschema
numpy
packaging
pandas
pip-review
pymisp
pyparsing
pyrsistent
PyYAML
python-dateutil
pytz
requests
six
urllib3
wrapt
zipp